+++ src/components/common/sidebar/projectList/projectList.style.ts
@@ -1,0 +1,20 @@
+import styled from 'styled-components';
+
+export const Projects = styled.div`
+  margin-top: 80px;
+  padding: 0 10.938%;
+  padding-bottom: 20px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  cursor: pointer;
+`;
+
+export const ProjectText = styled.p`
+  font-size: 19.2px;
+  font-weight: 500;
+  line-height: 28.8px;
+  letter-spacing: 0.02em;
+  text-align: left;
+  color: #d6deec;
+`;
