+++ src/components/common/sidebar/projectList/projectList.tsx
@@ -1,0 +1,31 @@
+import { useDispatch } from '@/hooks/common/useCustomRedux.ts';
+import useProjectList from '@/hooks/sidebar/sidebar';
+
+import { MODAL_TYPES } from '@/components/common/modalProvider/modalProvider.tsx';
+import * as S from '@/components/common/sidebar/projectList/projectList.style';
+
+import ProjectItem from '../projectItem/ProjectItem';
+
+import Plus from '@/assets/icons/add.svg?react';
+import { openModal } from '@/slices/modalSlice.ts';
+
+export default function ProjectList() {
+  const modalDispatch = useDispatch();
+  const { useGetProjectList } = useProjectList();
+  const { data: projectList } = useGetProjectList;
+  const projects = projectList?.result.projectList || [];
+
+  return (
+    <>
+      <S.Projects className="menu">
+        <S.ProjectText>Projects</S.ProjectText>
+        <Plus
+          onClick={() => {
+            modalDispatch(openModal({ modalType: MODAL_TYPES.CreateProjectModal, modalProps: { projectLength: projectList?.result.projectList.length } }));
+          }}
+        />
+      </S.Projects>
+      <ProjectItem projects={projects} />
+    </>
+  );
+}
