+++ src/components/common/sidebar/sidebar.tsx
@@ -1,155 +1,30 @@
-import { useEffect, useRef, useState } from 'react';
+import { useRef } from 'react';
 
-import useProjectList from '@/hooks/sidebar/sidebar';
+import { useDispatch } from '@/hooks/common/useCustomRedux.ts';
 
-import Profile from '@/components/common/profile/profile';
+import { MODAL_TYPES } from '@/components/common/modalProvider/modalProvider.tsx';
 import * as S from '@/components/common/sidebar/sidebar.style';
 
-import LogoutModal from './logtoutModal/logoutModal';
-import ProjectModal from './projectModal/projectModal';
+import SideBarProfile from './profile/sideBarProfile';
+import ProjectList from './projectList/projectList';
 
-import Plus from '@/assets/icons/add.svg?react';
-import ArrowDown from '@/assets/icons/arrow_down.svg?react';
-import ArrowRight from '@/assets/icons/arrow_right.svg?react';
-import ArrowUp from '@/assets/icons/arrow_up.svg?react';
-import SenarioLogo from '@/assets/icons/file_branch.svg?react';
-import DashboardLogo from '@/assets/icons/grid.svg?react';
-import InformationLogo from '@/assets/icons/info.svg?react';
-import Logo from '@/assets/icons/logo.svg?react';
 import Out from '@/assets/icons/logout.svg?react';
-
-type TUserProfile = {
-  id: number;
-  name: string;
-  profileImg: string;
-};
+import { openModal } from '@/slices/modalSlice.ts';
 
 export default function Sidebar() {
-  const { useGetProjectList, useGetSidebarUserInfo } = useProjectList();
-  const { data: projectList } = useGetProjectList;
-  const { data: userInfo } = useGetSidebarUserInfo;
-  const projects = projectList?.result.projectList || [];
-  const nickname = userInfo?.result.nickname;
-  const profile = userInfo?.result.profileImage;
-  const userProfile: TUserProfile = {
-    id: 1,
-    name: nickname || '',
-    profileImg: profile || '',
-  };
-
-  const [menuStates, setMenuStates] = useState<boolean[]>(new Array(projects.length).fill(false));
+  const modalDispatch = useDispatch();
   const sidebarRef = useRef<HTMLDivElement | null>(null);
-  const [modalShow, setModalShow] = useState(false);
-  const [logoutModal, setLogoutModal] = useState(false);
-  useEffect(() => {
-    if (projects.length !== menuStates.length) {
-      setMenuStates(new Array(projects.length).fill(false));
-    }
-  }, [projects, menuStates.length]);
-
-  const showModal = () => {
-    setModalShow(true);
-  };
-  const hideModal = () => {
-    setModalShow(false);
-  };
-  const logoutShow = () => {
-    setLogoutModal(true);
-  };
-  const logoutHide = () => {
-    setLogoutModal(false);
-  };
-  const toggleMenu = (index: number) => {
-    setMenuStates((prevStates) => {
-      return prevStates.map((state, i) => {
-        if (i === index) {
-          return !state;
-        }
-        return state;
-      });
-    });
-  };
-
   return (
     <S.SideBar ref={sidebarRef}>
       <S.LogoutBox>
         <S.ProjectBox>
-          <S.Container>
-            <S.StyledNavLink to={`/project`}>
-              <Logo width={25.6} height={25.6} />
-            </S.StyledNavLink>
-            <S.StyledNavLink to={`/mypage`}>
-              <S.Profile>
-                <S.SemiBox>
-                  <S.ProfileWrapper className="show project">
-                    <Profile profileImg={profile || ''} />
-                  </S.ProfileWrapper>
-                  <S.ProfileName className="menu">{userProfile.name}</S.ProfileName>
-                </S.SemiBox>
-                <ArrowRight className="menu" />
-              </S.Profile>
-            </S.StyledNavLink>
-          </S.Container>
-
-          <S.Projects className="menu">
-            <S.ProjectText>Projects</S.ProjectText>
-            <Plus
-              onClick={() => {
-                showModal();
-              }}
-            />
-            {modalShow && <ProjectModal projectLength={projectList?.result.projectList.length} onClose={hideModal} />}
-          </S.Projects>
-          <S.ProjectList>
-            <div>
-              {projects &&
-                projects.map((project, index) => (
-                  <div key={index}>
-                    <S.Project onClick={() => toggleMenu(index)}>
-                      <S.SemiBox>
-                        <S.ProjectProfile className="show content">
-                          <Profile profileImg={project.projectImage || ''} />
-                        </S.ProjectProfile>
-                        <S.ProjectName className="menu">{project.projectName}</S.ProjectName>
-                      </S.SemiBox>
-                      {menuStates[index] ? <ArrowUp className="menu" /> : <ArrowDown className="menu" />}
-                    </S.Project>
-                    <S.ProjectContents $isOpen={menuStates[index]} className="menu">
-                      <S.StyledNavLink to={`/project/dashboard/${project.projectId}`}>
-                        {({ isActive }) => (
-                          <S.ProjectContent $isActive={isActive}>
-                            <DashboardLogo width={19.2} height={19.2} />
-                            <S.ProjectContentName>Dashboard</S.ProjectContentName>
-                          </S.ProjectContent>
-                        )}
-                      </S.StyledNavLink>
-                      <S.StyledNavLink to={`/project/information/${project.projectId}`}>
-                        {({ isActive }) => (
-                          <S.ProjectContent $isActive={isActive}>
-                            <InformationLogo width={19.2} height={19.2} />
-                            <S.ProjectContentName>Information</S.ProjectContentName>
-                          </S.ProjectContent>
-                        )}
-                      </S.StyledNavLink>
-                      <S.StyledNavLink to={`/project/scenario/${project.projectId}`}>
-                        {({ isActive }) => (
-                          <S.ProjectContent $isActive={isActive}>
-                            <SenarioLogo width={19.2} height={19.2} />
-                            <S.ProjectContentName>Scenario</S.ProjectContentName>
-                          </S.ProjectContent>
-                        )}
-                      </S.StyledNavLink>
-                    </S.ProjectContents>
-                  </div>
-                ))}
-            </div>
-          </S.ProjectList>
+          <SideBarProfile />
+          <ProjectList />
         </S.ProjectBox>
-        <S.Logout onClick={logoutShow}>
+        <S.Logout onClick={() => modalDispatch(openModal({ modalType: MODAL_TYPES.LogoutModal }))}>
           <p className="menu">Logout</p>
           <Out />
         </S.Logout>
-        {logoutModal && <LogoutModal onClose={logoutHide} />}
       </S.LogoutBox>
     </S.SideBar>
   );
